requirements.txt
Flask
Flask-Admin
Flask-SQLAlchemy
flask-login
python-dotenv      # для .env-конфига
python-telegram-bot